updateded the search profile by samuel
